pyproject.toml
[project]
name = "job_skills"
version = "0.1.0"
description = "job_skills using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "wikipedia",
    "requests",
    "pandas",
    "python-dotenv",
]

[project.scripts]
job_skills = "job_skills.main:run"
run_crew = "job_skills.main:run"
train = "job_skills.main:train"
replay = "job_skills.main:replay"
test = "job_skills.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
use_active_env = true